Improved Camera modes
The camera can now switch modes between an overhead cam and a follow cam. Both cameras work but switch at incorrect points. I am unsure of how to fix this for the time being but I believe the problem is related to when the dice stops moving in some way.